Fix gnosis integration + refactor
